distro: Discard external-toolchain support
Really it seemed a good idea in the first place, but then turned to be too
problematic. First of all the compatibility with crosstools-NG still isn't
full and needs much efforts to make things fully working. Secondly, moving
to newer yoctoversion causes much pain, since the layer is supported just
by one person kergoth. Finally, we decided the natively built toolchain
is going to be used in the SDK, so we don't really need the externally
built one anymore. So just discard any meta-external-toolchainconfigs from
the recovery layer.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>