yocto: Move to Yocto 2.5 Sumo version with meta-oe layer
Last release of Yocto was named to rocko, but now 2.5 got released so the
BSP and recovery layers need to be made being compatible with it. First of all
due to glibc being updated to newer version a patch for older gcc compatibility
should be altered.

Another aspect of this patch are alterations connected with meta-openembedded/meta-oe
layer dependency. Since most of the benchmark and test applications used before
were part of that layer, we now either only need to a bbappend files with custom
configs or just discard the corresponding recipes from meta-recovery layer.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>