distro: Add micropython layer and recipe support
Miscropython is an embedded version of python created specifically for
devices with small amount of RAM and flash-memory with minimum functionality
deviation from the original python. As a result it is an ideal candidate for
recovery image, which reside in a just 16Mb flash.

Alas the newest micropython version isn't easy buildable with last libffi. We'll
have to disable the library support for now.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>