u-boot: Use u-boot default instead of custom recipe
New u-boot recipe is created based on the default one placed at the
common meta layer. It works in exactly the same way, but downloads
sources from our repo and cleans the deploy directory up as we
don't need that much bits and links there. So just fix a U-boot
preferred version to be the one from our repo.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>