distro: Add pulse-cntr app recipe
T-platforms MRBT1 mazannine board provides a functionality of pulse
counters attached to each of four BMC GPIOs. In order to simplify
the test process, we created an app called pulse-cntr, which provides
a simple command-line interface to manually send/recv commands
to the counter. In particular a user can read/write:
- pulse polarity
- debounce timer
- counter state
for each BMC pin.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>