distro: Discard kernel-base runtime dependency
The kernel vmlinux image gets to be installed in case if any kernel
module is built and installed to the recovery rootfs. It dramatically
expands the recovery image. Since we really don't need it being
embedded into the recovery rootfs tree in any case we can freely remove
the dependency here for the whole distro.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>