initscripts: Add load console fonts script
We need a initscript to automatically load console font on boot stage
in order to improve the recovery mode picture on graphic monitors.
This commit provides a script for doing this. If a machine-specific
config does not define RECOVERY_FONT variable, the script is not
installed to the rootfs and shall not be run at the system startup.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>