kernel: Create oland subset of radeon firmware
bn1bt1 is equipped with AMD Oland Radeon GPU which requires a firmware
for normal operation. This firmware is delivered as a part of
linux-firmware package. It will be installed to the recovery image
to have the Oland radeon graphics support in recovery mode. But we'll
install only few binaries required for Oland so no to waste the image
space with unused binaries.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>