distro: Add serkbd package recipe
serkbd is a utility providing support for PS/2-like serio over serial
tty line. It is required to enable the ps/2 keyboard on bn1bt1 as an
event-device by switching the serial port specific configs.

Since the application is very simple we didn't create a dedicated repo
for it. It'source code is added directly to the recipe together with
the corresponding install-script.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>