distro: Add bn1bt1 board config
bn1bt1 is a laptop machine created by T-platforms and based on
Baikal-T1 SoC. The patch adds the recovery distro config for it.
We install ttm and radeon kernel modules to have a support of the
AMD video chip installed on the circuit, enable four video getties,
setup fonts on the recovery system startup, iniialize bootconf-based
and bootrecovery kernel bootup procedure.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>