distro: Set valid USB interfaces in U-boot bootmenu
Indeed even though the USB-related boot items stated to load to some
sdX rootfs, they load the kernel, initramfs and dtb from a single USB
interface, which in most of the cases wasn't the same as the rootfs
device. In order to fix this at least with small limitation lets
set device:partition pairs of U-boot USB boot items corresponding to
the Linux block devices as if there is no other devices rather than
USB ones. In this case it's at least much easier to undertand the boot
scenarios.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>