distro: Add linkloop utility recipe
This is a very useful utility to perform a link-level (layer 2 of OSI)
ping either from one host of another or a loopbacked host. In first case
a server-like application called linkloop_reply is supposed to be executed
on the being pinged host. We need these utilities in the framework of
TP-test ethernet testcases.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>