distro: Add TP-test framework recipe
T-platforms Test framework is a special program to perform a functional
tests of T-platforms embedded products. In particular it consist of
the core application, which is responsible for the test sequenced
executaion, and of testcases and scenarios, which are dsigned to test
the corresponding products hardware/software peculirities. The framework
is split into the Yocto packages in accordance with the test subsystems
and scenario, so the recovery image would contain only a dedicated set
of testcases.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>