smstools3: Use ttyUSB7 for AT-commands by default
The only target, which is using this utility is BC3BT1-MRBT1 one, where
a Sierra Wireless em7430 modem is installed. It turns out the AT-commands
interface is exposed in the /dev/ttyUSB7 device, while DM and GNSS are
accessible via the /dev/ttyUSB5 and /dev/ttyUSB6 files. Lets remove the
rest of interfaces from the smsd config file, since they can't be used
for actual SMS transmission.

Signed-off-by: Sergey Semin <fancer.lancer@gmail.com>